+++ src/components/homepage/SectionMusic.js
@@ -7,9 +7,9 @@ export function SectionMusic() {
   const featuredMusic = info.music[randomIndex];
 
   return (
-    <div className="flex-col align-items pb-10">
-      <h1 className="font-light text-5xl flex justify-center pb-10">Music</h1>
-      <div className="flex justify-center">
+    <div>
+      <h1 className="font-light text-5xl flex justify-center my-10">Music</h1>
+      <div className="flex justify-center my-5">
         <MusicCard
           id={featuredMusic.id}
           title={featuredMusic.title}
@@ -22,10 +22,10 @@ export function SectionMusic() {
           }
         />
       </div>
-      <div className="px-36 flex justify-center mt-5">
+      <div className="flex justify-center mt-10">
         <Link to="/music">
           <div>
-            <span className="text-md bg-[#444] text-[#eee] p-3 rounded-md hover:bg-[#555] duration-300">
+            <span className="text-md bg-[#444] text-white p-3 rounded-md hover:bg-[#555] duration-300">
               All Music
             </span>
           </div>