+++ src/components/homepage/SectionProjects.js
@@ -6,11 +6,9 @@ export function SectionProjects() {
   const featuredProjects = projects.slice(0, 3);
 
   return (
-    <div className="flex-col align-items py-10">
-      <h1 className="font-light text-5xl flex justify-center pb-10">
-        Projects
-      </h1>
-      <div className="grid grid-cols-3 gap-20 px-36 py-10">
+    <div className="py-5">
+      <h1 className="font-light text-5xl flex justify-center py-5">Projects</h1>
+      <div className="md:grid md:grid-cols-3 flex flex-col md:flex-col gap-6 md:gap-20 px-6 md:px-12 py-6 md:py-10">
         {featuredProjects.map((project) => (
           <HomeProjectCard
             name={project.name}
@@ -21,12 +19,12 @@ export function SectionProjects() {
           </HomeProjectCard>
         ))}
       </div>
-      <div className="px-36 text-blue-700 font-light text-lg">
+      <div className="px-6 md:px-12 text-blue-700 font-light text-md">
         <Link to="/projects">
-          <div>
+          <h1>
             <span className="underline">All projects</span> {">"}
             {">"}
-          </div>
+          </h1>
         </Link>
       </div>
     </div>