+++ src/pages/Projects.js
@@ -3,9 +3,9 @@ import { projects } from "../utils";
 
 export function Projects() {
   return (
-    <div className="flex flex-col items-center py-10 bg-gradient-to-b from-blue-200 via-white to-sky-100">
-      <h1 className="text-5xl mb-5 ">Projects</h1>
-      <div className="grid grid-cols-2 flex justify-items-center gap-10 p-10">
+    <div className="flex flex-col items-center py-5 bg-gradient-to-b from-blue-200 via-white to-red-50">
+      <h1 className="text-5xl my-5">Projects</h1>
+      <div className="flex flex-col items-center md:grid md:grid-cols-2 gap-10 p-10">
         {projects.map((item) => (
           <ProjectCard
             name={item.name}